Merge Dev

+++ src/www/assets/js/settings.js
@@ -1,6 +1,7 @@
 
 import { translate, updateTranslate, initI18n, getLanguageFromLocalStorage } from './i18n.js';
 import { getThemeFromLocalStorage, setThemeInLocalStorage } from './theme.js';
+// import { Toast } from 'bootstrap';
 
 // import { flatpickr } from '../flatpickr/js/flatpickr.js';
 
@@ -24,6 +25,55 @@ let lang = "en-US"; // Default language
  */
 
 document.addEventListener("DOMContentLoaded", async () => {
+
+    // Funzione per mostrare un toast Bootstrap
+    function showToast(message, options = {}) {
+        // Cerca o crea il contenitore dei toast
+        let toastContainer = document.getElementById('toast-container');
+        if (!toastContainer) {
+            toastContainer = document.createElement('div');
+            toastContainer.id = 'toast-container';
+            toastContainer.className = 'toast-container position-fixed bottom-0 end-0 p-3';
+            document.body.appendChild(toastContainer);
+        }
+
+        // Crea il markup del toast
+        const toastElem = document.createElement('div');
+        toastElem.className = 'toast align-items-center text-bg-primary border-0';
+        toastElem.setAttribute('role', 'alert');
+        toastElem.setAttribute('aria-live', 'assertive');
+        toastElem.setAttribute('aria-atomic', 'true');
+        toastElem.innerHTML = `
+            <div class="d-flex">
+                <div class="toast-body">
+                    ${message}
+                </div>
+                <button type="button" class="btn-close btn-close-white me-2 m-auto" data-bs-dismiss="toast" aria-label="Close"></button>
+            </div>
+        `;
+
+        toastContainer.appendChild(toastElem);
+
+        
+        // Opzioni di default
+        const toastOptions = {
+            delay: options.delay || 3000,
+            autohide: options.autohide !== undefined ? options.autohide : true
+        };
+        // $('.toast').toast(toastOptions);
+
+        // Inizializza e mostra il toast
+        // const toast = new Toast(toastElem, toastOptions);
+        // toast.show();
+
+        // Rimuovi il toast dal DOM quando viene nascosto
+        toastElem.addEventListener('hidden.bs.toast', () => {
+            toastElem.remove();
+        });
+    }
+
+    // Esempio d'uso:
+    // showToast('Hello, this is a Bootstrap toast!');
 
     const themeItems = document.querySelectorAll('.theme-item');
 
@@ -116,7 +166,7 @@ document.addEventListener("DOMContentLoaded", async () => {
             }
         } catch (error) {
             console.error("Error adding printer:", error);
-            alert(translate('Error adding printer: ') + error.message);
+            showToast(translate('Error adding printer: ') + error.message);
         }
     };
 
@@ -134,14 +184,15 @@ document.addEventListener("DOMContentLoaded", async () => {
             
             console.log("Delete printer result:", response);
             if (response) {
-                alert(translate('Printer deleted successfully'));
+                // alert(translate('Printer deleted successfully'));
+                showToast('Printer deleted successfully');
                 await fillPrinterTable(); // Refresh the table after deletion
             } else {
-                alert(translate('Error deleting printer: ') + response);
+                showToast(translate('Error deleting printer: ') + response);
             }
         } catch (error) {
             console.error("Error deleting printer:", error);
-            alert(translate('Error deleting printer: ') + error.message);
+            showToast(translate('Error deleting printer: ') + error.message);
         }
     };
 
@@ -187,14 +238,17 @@ document.addEventListener("DOMContentLoaded", async () => {
             }
             console.log("Save result:", response);
             if (response) {
-                alert(translate('Printers saved successfully'));
+                // alert(translate('Printers saved successfully'));
+                showToast(translate('Printers saved successfully'));
                 await fillPrinterTable(); // Refresh the table after saving
             } else {
-                alert(translate('Error saving printers: ') + response);
+                // alert(translate('Error saving printers: ') + response);
+                showToast(translate('Error saving printers: ') + response);
             }
         } catch (error) {
             console.error("Error saving printers:", error);
-            alert(translate('Error saving printers: ') + error.message);
+            // alert(translate('Error saving printers: ') + error.message);
+            showToast(translate('Error saving printers: ') + error.message);
         }
     };
 
@@ -219,9 +273,14 @@ document.addEventListener("DOMContentLoaded", async () => {
                 </td>
                 <td><input type="text" class="form-control" id="printer-description-${printer.key}" value="${printer.description}"></td>
                 <td>
-                    <button class="btn btn-danger btn-sm btn-delete" id="btn-delete-${printer.key}" data-printer="${printer.key}" title="Delete">
-                        <i class="bi bi-trash"></i>
-                    </button>
+                    <div class="d-flex justify-content-center align-items-center">
+                        <button class="btn btn-primary btn-sm me-2" id="btn-test-${printer.key}" data-printer="${printer.key}" title="TEST">
+                            <i class="bi bi-printer-fill"></i>
+                        </button>
+                        <button class="btn btn-danger btn-sm btn-delete" id="btn-delete-${printer.key}" data-printer="${printer.key}" title="Delete">
+                            <i class="bi bi-trash"></i>
+                        </button>
+                    </div>
                 </td>
             `;
             printerTableBody.appendChild(row);
@@ -245,39 +304,23 @@ document.addEventListener("DOMContentLoaded", async () => {
                         <td><input type="checkbox" class="form-check-input printer-active-checkbox" id="printer-active-${newPrinterKey}"></td>
                         <td><input type="text" class="form-control" id="printer-description-${newPrinterKey}" value=""></td>
                         <td>
-                            <button class="btn btn-danger btn-sm btn-delete" id="btn-delete-${newPrinterKey}" data-printer="${newPrinterKey}" title="Delete">
-                                <i class="bi bi-trash"></i>
-                            </button> 
+                            <div class="d-flex justify-content-center align-items-center">
+                                <button class="btn btn-primary btn-sm me-2" id="btn-test-${newPrinterKey}" data-printer="${newPrinterKey}" title="TEST">
+                                    <i class="bi bi-printer-fill"></i>
+                                </button>
+                                <button class="btn btn-danger btn-sm btn-delete" id="btn-delete-${newPrinterKey}" data-printer="${newPrinterKey}" title="Delete">
+                                    <i class="bi bi-trash"></i>
+                                </button>
+                            </div>
                         </td>
                     `;
                     const printerTableBody = document.getElementById('printers-table-body');
                     if (printerTableBody) {
                         printerTableBody.appendChild(newRow);
                     }
-                    // Add event listeners for the new row's buttons
-                    const deleteButton = newRow.querySelector('.btn-delete');
-                    if (deleteButton) {
-                        deleteButton.addEventListener('click', async () => {
-                            const printerKey = deleteButton.dataset.printer;
-                            if (printerKey) {
-                                await deletePrinter(printerKey);
-                            }
-                        });
-                    }
                 }
             });
         }
-
-        const deleteButtons = document.querySelectorAll('.btn-delete');
-        deleteButtons.forEach(button => {
-            button.addEventListener('click', async () => {
-                const printerKey = button.dataset.printer;
-                if (printerKey) {
-                    console.log("Deleting printer:", printerKey);
-                    await deletePrinter(printerKey);
-                }
-            });
-        });
 
         const saveButtons = document.querySelectorAll('#save-printers-btn');
         saveButtons.forEach(button => {
@@ -294,12 +337,40 @@ document.addEventListener("DOMContentLoaded", async () => {
                 const printerKey = button.dataset.printer;
                 if (printerKey) {
                     console.log("Deleting printer:", printerKey);
+                    if (!confirm(translate('Are you sure you want to delete this printer?'))) {
+                        return;
+                    }
                     await deletePrinter(printerKey);
+                }
+            });
+        });
+
+        const testButtons = document.querySelectorAll('[id^="btn-test-"]');
+        testButtons.forEach(button => {
+            button.addEventListener('click', async () => {
+                const printerKey = button.dataset.printer;
+                if (printerKey) {
+                    console.log("Testing printer:", printerKey);
+                    const response = await fetch(`/api/printers/test/${printerKey}`, {
+                        method: 'POST',
+                        headers: {
+                            'Content-Type': 'application/json'
+                        }
+                    });
+                    if (!response.ok) {
+                        throw new Error(`HTTP error! status: ${response.status}`);
+                    }
+                    if(response) {
+                        alert(translate('Printer test successful: ') + printerKey);
+                    }
+                    // Implement the test functionality here
+                    // alert(translate('Testing printer: ') + printerKey);
                 }
             });
         });
     };
     
+    // -- Initialize the language and theme
 
     lang = await initI18n();
     applyThemeFromLocalStorage();
